Holding variable constant and generalizing map creation
Made the intitial conditions in calcReddening.py based off of distance and not index.
This means when fitting maps with different pixel sizes that the predictions, there is no discontinuity.
Also began holding the paramter, B, constant in minimizer.py.
The parameter B reduces the amount of fitting parameters as well as stabilizing the fitting process.